+++ App_Data/jobs/continuous/makeTransactionExpired/run.ts
@@ -1,12 +1,15 @@
 /**
- * タスクリトライ
+ * 取引期限監視
  *
  * @ignore
  */
 
 import * as sskts from '@motionpicture/sskts-domain';
+import * as createDebug from 'debug';
 
 import mongooseConnectionOptions from '../../../../mongooseConnectionOptions';
+
+const debug = createDebug('sskts-jobs:*');
 
 sskts.mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 
@@ -14,8 +17,6 @@ let count = 0;
 
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
 const INTERVAL_MILLISECONDS = 1000;
-const RETRY_INTERVAL_MINUTES = 10;
-const taskAdapter = sskts.adapter.task(sskts.mongoose.connection);
 
 setInterval(
     async () => {
@@ -26,7 +27,8 @@ setInterval(
         count += 1;
 
         try {
-            await sskts.service.task.retry(RETRY_INTERVAL_MINUTES)(taskAdapter);
+            debug('transaction expiring...');
+            await sskts.service.transaction.makeExpired()(sskts.adapter.transaction(sskts.mongoose.connection));
         } catch (error) {
             console.error(error.message);
         }