+++ App_Data/jobs/continuous/sendEmailNotification/run.js
@@ -14,21 +14,19 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const sskts = require("@motionpicture/sskts-domain");
-const mongoose = require("mongoose");
 const mongooseConnectionOptions_1 = require("../../../../mongooseConnectionOptions");
-mongoose.Promise = global.Promise;
-mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions_1.default);
+sskts.mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions_1.default);
 let count = 0;
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
 const INTERVAL_MILLISECONDS = 500;
-const taskAdapter = sskts.adapter.task(mongoose.connection);
+const taskAdapter = sskts.adapter.task(sskts.mongoose.connection);
 setInterval(() => __awaiter(this, void 0, void 0, function* () {
     if (count > MAX_NUBMER_OF_PARALLEL_TASKS) {
         return;
     }
     count += 1;
     try {
-        yield sskts.service.task.executeByName(sskts.factory.taskName.SendEmailNotification)(taskAdapter, mongoose.connection);
+        yield sskts.service.task.executeByName(sskts.factory.taskName.SendEmailNotification)(taskAdapter, sskts.mongoose.connection);
     }
     catch (error) {
         console.error(error.message);