+++ App_Data/jobs/continuous/sendEmailNotification/run.ts
@@ -5,18 +5,16 @@
  */
 
 import * as sskts from '@motionpicture/sskts-domain';
-import * as mongoose from 'mongoose';
 
 import mongooseConnectionOptions from '../../../../mongooseConnectionOptions';
 
-(<any>mongoose).Promise = global.Promise;
-mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
+sskts.mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 
 let count = 0;
 
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
 const INTERVAL_MILLISECONDS = 500;
-const taskAdapter = sskts.adapter.task(mongoose.connection);
+const taskAdapter = sskts.adapter.task(sskts.mongoose.connection);
 
 setInterval(
     async () => {
@@ -29,7 +27,7 @@ setInterval(
         try {
             await sskts.service.task.executeByName(
                 sskts.factory.taskName.SendEmailNotification
-            )(taskAdapter, mongoose.connection);
+            )(taskAdapter, sskts.mongoose.connection);
         } catch (error) {
             console.error(error.message);
         }