+++ App_Data/jobs/continuous/settleGMOAuthorization/run.ts
@@ -6,20 +6,18 @@
 
 import * as sskts from '@motionpicture/sskts-domain';
 import * as createDebug from 'debug';
-import * as mongoose from 'mongoose';
 
 import mongooseConnectionOptions from '../../../../mongooseConnectionOptions';
 
 const debug = createDebug('sskts-jobs:bin:watchSettleGMOAuthorizationQueue');
 
-(<any>mongoose).Promise = global.Promise;
-mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
+sskts.mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 
 let count = 0;
 
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
 const INTERVAL_MILLISECONDS = 1000;
-const taskAdapter = sskts.adapter.task(mongoose.connection);
+const taskAdapter = sskts.adapter.task(sskts.mongoose.connection);
 
 setInterval(
     async () => {
@@ -33,7 +31,7 @@ setInterval(
             debug('count:', count);
             await sskts.service.task.executeByName(
                 sskts.factory.taskName.SettleGMOAuthorization
-            )(taskAdapter, mongoose.connection);
+            )(taskAdapter, sskts.mongoose.connection);
         } catch (error) {
             console.error(error.message);
         }