+++ App_Data/jobs/continuous/settleMvtkAuthorization/run.js
@@ -16,15 +16,13 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 Object.defineProperty(exports, "__esModule", { value: true });
 const sskts = require("@motionpicture/sskts-domain");
 const createDebug = require("debug");
-const mongoose = require("mongoose");
 const mongooseConnectionOptions_1 = require("../../../../mongooseConnectionOptions");
 const debug = createDebug('sskts-jobs:bin:watchSettleMvtkAuthorizationQueue');
-mongoose.Promise = global.Promise;
-mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions_1.default);
+sskts.mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions_1.default);
 let count = 0;
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
 const INTERVAL_MILLISECONDS = 1000;
-const taskAdapter = sskts.adapter.task(mongoose.connection);
+const taskAdapter = sskts.adapter.task(sskts.mongoose.connection);
 setInterval(() => __awaiter(this, void 0, void 0, function* () {
     if (count > MAX_NUBMER_OF_PARALLEL_TASKS) {
         return;
@@ -32,7 +30,7 @@ setInterval(() => __awaiter(this, void 0, void 0, function* () {
     count += 1;
     try {
         debug('count:', count);
-        yield sskts.service.task.executeByName(sskts.factory.taskName.SettleMvtkAuthorization)(taskAdapter, mongoose.connection);
+        yield sskts.service.task.executeByName(sskts.factory.taskName.SettleMvtkAuthorization)(taskAdapter, sskts.mongoose.connection);
     }
     catch (error) {
         console.error(error.message);