+++ App_Data/jobs/continuous/watchExpiredTransaction/run.ts
@@ -1,37 +1,41 @@
 /**
- * タスクリトライ
+ * 期限切れ取引監視
  *
  * @ignore
  */
 
 import * as sskts from '@motionpicture/sskts-domain';
+import * as createDebug from 'debug';
 
 import mongooseConnectionOptions from '../../../../mongooseConnectionOptions';
 
+const debug = createDebug('sskts-jobs:*');
+
 sskts.mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 
-let count = 0;
+let countExecute = 0;
 
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
-const INTERVAL_MILLISECONDS = 1000;
-const RETRY_INTERVAL_MINUTES = 10;
+const INTERVAL_MILLISECONDS = 500;
 const taskAdapter = sskts.adapter.task(sskts.mongoose.connection);
+const transactionAdapter = sskts.adapter.transaction(sskts.mongoose.connection);
 
 setInterval(
     async () => {
-        if (count > MAX_NUBMER_OF_PARALLEL_TASKS) {
+        if (countExecute > MAX_NUBMER_OF_PARALLEL_TASKS) {
             return;
         }
 
-        count += 1;
+        countExecute += 1;
 
         try {
-            await sskts.service.task.retry(RETRY_INTERVAL_MINUTES)(taskAdapter);
+            debug('exporting queues...');
+            await sskts.service.transaction.exportTasks(sskts.factory.transactionStatus.EXPIRED)(taskAdapter, transactionAdapter);
         } catch (error) {
             console.error(error.message);
         }
 
-        count -= 1;
+        countExecute -= 1;
     },
     INTERVAL_MILLISECONDS
 );